WIP adding php tests
Start of writing tests in php itself

Adding first PHP test, debugging test failure

why am I getting a different error than when accessing
this API through the real application?

Problem was I added an error handler that would capture everything

Including "errors" of type "notice", which don't normally stop the
script, and in my error handler I was echoing information about
the error/notice to the request output, which I want to be able to
capture and parse as JSON in most cases.

Overall putting test results into the overall browser requests for
the "runTests" api, while wanting to capture output of individual
scripts and assert on them is a bit unergenomic, but I think I might
have worked out the bulk of the problems, the current test file looks
pretty clean and the extra stuff is hidden in 'utilities.php' and the runner.

Solution for now, just make the error handler put info in the logs,
but in the longer run it might be nice to surface this kind of info
in the test results to make it more obvious.

Utility function for checking a sql table from a test